# backend/service.core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.HomePageReport:
    properties:
      all_time:
        $ref: '#/definitions/entity.HomePageReportItem'
        type: object
      daily:
        $ref: '#/definitions/entity.HomePageReportItem'
        type: object
      generated_at:
        type: string
      monthly:
        $ref: '#/definitions/entity.HomePageReportItem'
        type: object
      quarterly:
        $ref: '#/definitions/entity.HomePageReportItem'
        type: object
      weekly:
        $ref: '#/definitions/entity.HomePageReportItem'
        type: object
      yearly:
        $ref: '#/definitions/entity.HomePageReportItem'
        type: object
    type: object
  entity.HomePageReportItem:
    properties:
      most_ordered_category:
        $ref: '#/definitions/entity.MenuItemCategory'
        type: object
      most_ordered_item:
        $ref: '#/definitions/entity.MenuItem'
        type: object
      most_ordered_item_quantity:
        type: integer
      total_revenue:
        type: number
    type: object
  entity.MenuItem:
    properties:
      allergy:
        type: string
      category_id:
        type: string
      description:
        type: string
      file:
        type: string
      id:
        type: string
      is_menu:
        type: boolean
      is_special:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  entity.MenuItemCategory:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  entity.Restaurant:
    properties:
      businessHours:
        type: string
      cuisine:
        type: string
      email:
        type: string
      file:
        type: string
      id:
        example: res_abc123d
        type: string
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      type:
        type: string
      website:
        type: string
    type: object
...
  errors.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  restaurants.CreateRestaurantRequest:
    properties:
      businessHours:
        type: string
      cuisine:
        type: string
      email:
        type: string
      file:
        type: string
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      type:
        type: string
      website:
        type: string
    type: object
  restaurants.ItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.Restaurant'
        type: object
    type: object
  restaurants.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.Restaurant'
        type: array
    type: object
  restaurants.UpdateRestaurantRequest:
    $ref: '#/definitions/restaurants.CreateRestaurantRequest'
...
host: localhost:5000
info:
  contact: {}
  description: Restaurants and customers and other great things!
  license: {}
  title: COMP3900 - JAMAR
  version: "0.1"
paths:
...
  /recommended-restaurants:
    get:
      description: By default returns all restaurants recommended based on this user's previous order history. For detailed explaination, please read our software report
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/restaurants.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a list of restaurants, recommended for this specific user
      tags:
      - Restaurants
...
  /restaurants:
    get:
      description: By default returns all restaurants in the system but managers can use the `mine=true` query param to only return restaurants they manage
      parameters:
      - description: If it's true, only return restaurants managed by the user
        in: query
        name: mine
        type: boolean
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/restaurants.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a list of restaurants
      tags:
      - Restaurants
    post:
      parameters:
      - description: The new restaurant's details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/restaurants.CreateRestaurantRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/restaurants.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Creates a new restaurant attached to the user's account
      tags:
      - Restaurants
  /restaurants/{res_id}:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/entity.Restaurant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns the details for the specified restaurant
      tags:
      - Restaurants
    put:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The restaurant's updated details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/restaurants.UpdateRestaurantRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/entity.Restaurant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Updates a restaurant's details
      tags:
      - Restaurants
...
  /restaurants/{res_id}/report:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: type of report
        in: query
        name: type
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/entity.HomePageReport'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns the details for the specified restaurant. Currently only support "home" as a type.
      tags:
      - Restaurants
...
swagger: "2.0"